+++ breeze-data/src/main/kotlin/com/windea/breezeframework/data/extensions/CoreExtensions.kt
@@ -7,7 +7,6 @@ import com.windea.breezeframework.core.extensions.*
 import com.windea.breezeframework.data.enums.*
 import com.windea.breezeframework.data.serializers.*
 import com.windea.breezeframework.reflect.extensions.java.*
-import mu.*
 import java.io.*
 
 //REGION serialize extensions