+++ breeze-logger/src/main/kotlin/com/windea/breezeframework/logger/Extensions.kt
@@ -1,0 +1,5 @@
+package com.windea.breezeframework.logger
+
+import com.windea.breezeframework.logger.internal.*
+
+val logger: Logger = ColorfulLogger