+++ breeze-logger/src/main/kotlin/com/windea/breezeframework/logger/LogLevel.kt
@@ -1,0 +1,13 @@
+package com.windea.breezeframework.logger
+
+enum class LogLevel(
+	val text: String
+) {
+	Off("OFF"),
+	Trace("TRACE"),
+	Debug("DEBUG"),
+	Info("INFO"),
+	Warn("WARN"),
+	Error("ERROR"),
+	Fatal("FATAL")
+}